backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pyjwt==2.9.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# PDF processing
PyMuPDF==1.24.11
pytesseract==0.3.13
Pillow==10.4.0

# AI / LLM
openai==1.50.2

# HTTP client
httpx==0.27.2

# Utils
python-multipart==0.0.12
python-dotenv==1.0.1
